+++ api/src/main/java/com/intuit/tank/vm/vmManager/JobRequest.java
@@ -55,7 +55,7 @@ public interface JobRequest extends Serializable {
     /**
      * @return the endRate
      */
-    public abstract int getEndRate();
+    public abstract double getEndRate();
 
     /**
      * @return the numAgents
@@ -96,7 +96,7 @@ public interface JobRequest extends Serializable {
     /**
      * @return the userIntervalIncrement
      */
-    public abstract int getUserIntervalIncrement();
+    public abstract double getUserIntervalIncrement();
 
     /**
      * @return the reportingMode
